Rename project structure and configuration properties from sureflap to surehub
