+++ public/index.html
@@ -7,7 +7,7 @@
             type="text/css"
             href="https://cdn.jsdelivr.net/npm/swagger-ui-dist@5/swagger-ui.min.css"
     />
-    <title>SureFlap API</title>
+    <title>SureHub API</title>
 </head>
 <body>
 <div id="swagger-ui"></div>
