+++ pyproject.toml
@@ -1,11 +1,11 @@
 [tool.poetry]
-name = "sureflap-api"
+name = "surehub-api"
 version = "4.2.7"
 license = "Apache-2.0"
-description = "SureFlap API is a simple, yet powerful RESTful API for products from Sure Petcare"
+description = "SureHub API is a simple, yet powerful RESTful API for products from Sure Petcare"
 authors = ["Fabian Eulitz <dev@sustineo.de>"]
 repository = "https://github.com/fabieu/sureflap-api"
-keywords = ["Sureflap", "Sure Petcare", "REST", "API"]
+keywords = ["SureHub", "Sure Petcare", "REST", "API"]
 
 [tool.poetry.dependencies]
 python = "^3.10,<4.0"
