+++ surehub_api/config.py
@@ -1,7 +1,7 @@
 from dynaconf import Dynaconf, Validator
 
 settings = Dynaconf(
-    envvar_prefix="SUREFLAP",
+    envvar_prefix="SUREHUB",
     settings_files=["settings.yaml", ".secrets.yaml"],
     validators=[
         Validator('email', 'password', must_exist=True),
