+++ surehub_api/modules/dashboard.py
@@ -3,9 +3,9 @@ import json
 import requests
 from fastapi import HTTPException
 
-from sureflap_api.config import settings
-from sureflap_api.models import surehub
-from sureflap_api.modules import auth
+from surehub_api.config import settings
+from surehub_api.models import surehub
+from surehub_api.modules import auth
 
 
 def get_dashboard() -> surehub.MeStart:
