+++ surehub_api/modules/timeline.py
@@ -4,8 +4,8 @@ import math
 import requests
 from fastapi import HTTPException
 
-from sureflap_api.config import settings
-from sureflap_api.modules import auth
+from surehub_api.config import settings
+from surehub_api.modules import auth
 
 
 def get_timeline_of_household(household_id: int) -> list:
